fix(vm): keep a string's pool slot at full width on the free list
The string collector pushed each freed slot onto its free list as a u16.
Once the pool passed 65536 strings, a freed slot such as 70000 went back
out as 4464, a slot a live string still held, and the next long string
written there overwrote it. A program holding 70000 strings through a
collection read back 61070 of them wrong.

The free list now holds u32 slot ids, the width the string value itself
carries.

+++ vm/src/data.rs
@@ -309,7 +309,7 @@ impl Data {
         string_pool: &mut StringPool,
         registers: &RegisterFile,
         recursion_stack: &RegisterFile,
-        free_strings: &mut Vec<u16>,
+        free_strings: &mut Vec<u32>,
         gc_string_threshold: &mut u32,
         gc: &mut GcScratch,
     ) -> Self {
